fix: Missing thresholds leading to visualization crashing
The recent release of Query Builder included multiple improvements, most importantly, it is now able to edit custom visualizations, as well as validate and sanitize the configuration of the custom visualization.

The schema defined in the `nr1.json` file establishes the shape of the configuration supported and understood by the visualization. Because visualizations can be used in many contexts, we can't make any assumptions about the configuration being passed to it - whether all fields will be included and whether they will be of a particular type.

Therefore, it's on the visualization to protect from missing values and to handle these scenarios gracefully. While in Query Builder, the validation layer is going to prevent widgets from being saved with incorrect types, therefore, while building and editing widgets through New Relic, we have to guard only against the missing values scenario.