fix: Improved handling of missing properties
**Summary**

We're improving the validation logic not only to look for the empty values such as `null` and `''`, but also for missing properties/values (`undefined`).

**More context**

Unless specified otherwise [^footnote1], all fields defined in the schema of the visualization are nullable. This becomes especially important in New Relic Dashboards, where visualizations can be added in numerous ways, i.e. programmatically, or through the New Relic UI. Therefore, in addition to checking for null values and empty strings, we should also be guarding against undefined values (missing properties).

This commit improves validation logic on the visualization's side to offer the same experience with null as well as missing values. It's especially important while editing the dashboard's visualizations since Query Builder would strip out all undesired and empty values (unless they are marked as desired, i.e. in enums [^footnote2]) to keep the configuration as small and up to date as possible.

[^footnote1]: fields can be marked as required in the schema of the visualization and won't be allowed to be persisted through the UI unless a valid value is provided
[^footnote2]: when defining enum fields in the visualization's schema we have to enumerate the allowed values. The usage of null is allowed in enum fields, and when used, it wouldn't be stripped out by Query Builder